get propagate_shape logic out of module.h
This logic never made sense to be defined in a user exposed file,
and never worked for any types but tensors. This moves them to
internal functions and updates onnx/tests to use it. It also makes
the functions work directly on graphs, since Modules will lose their
initial_ivalues property soon.

Test Plan: python test_jit.py
