+++ torch/csrc/jit/script/module.h
@@ -623,75 +623,6 @@ struct TORCH_API Module {
   mutable std::recursive_mutex find_method_guard_;
 };
 
-static void setInputTensorTypes(Graph& g, const Stack& stack) {
-  AT_ASSERT(stack.size() == g.inputs().size());
-  for (size_t i = 0; i < stack.size(); ++i) {
-    g.inputs().at(i)->setType(
-        DimensionedTensorType::create(stack.at(i).toTensor()));
-  }
-}
-
-inline std::shared_ptr<Graph> propagate_shapes(
-    Graph& graph,
-    const std::vector<at::Tensor>& inputs,
-    const std::vector<Slot>& initial_ivalues,
-    bool with_grad = false) {
-  auto retval = graph.copy();
-  Stack stack;
-  stack.reserve(inputs.size() + initial_ivalues.size());
-  for (const at::Tensor& i : inputs) {
-    stack.emplace_back(std::move(i));
-  }
-  for (const Slot& inp : initial_ivalues) {
-    stack.push_back(inp.value());
-  }
-  setInputTensorTypes(*retval, stack);
-  PropagateInputShapes(retval);
-  return retval;
-}
-
-inline std::shared_ptr<Graph> propagate_and_assign_input_and_output_shapes(
-    Graph& graph,
-    std::vector<at::Tensor> inputs,
-    const std::vector<Slot>& initial_ivalues,
-    std::vector<at::Tensor> outputs,
-    bool with_grad = false,
-    bool propagate = true) {
-  auto retval = graph.copy();
-  for (auto inp : initial_ivalues) {
-    if (inp.value().isTensor()) {
-      inputs.push_back(inp.value().toTensor());
-    }
-  }
-  if (propagate) {
-    setInputTensorTypes(*retval, fmap<IValue>(inputs));
-    PropagateInputShapes(retval);
-  }
-  AT_ASSERT(retval->inputs().size() == inputs.size());
-  for (size_t i = 0; i < retval->inputs().size(); ++i) {
-    auto scalar_type = inputs[i].scalar_type();
-    auto sizes = inputs[i].sizes();
-    auto type =
-        torch::jit::CompleteTensorType::create(scalar_type, at::kCPU, sizes);
-    retval->inputs()[i]->setType(type);
-  }
-  at::ArrayRef<Value*> output_values = retval->outputs();
-  // patch this to still work if we are returning a tuple of multiple values
-  if (output_values.at(0)->type()->kind() == TupleType::Kind) {
-    AT_ASSERT(output_values.at(0)->node()->kind() == prim::TupleConstruct);
-    output_values = output_values.at(0)->node()->inputs();
-  }
-  AT_ASSERT(output_values.size() == outputs.size());
-  for (size_t i = 0; i < retval->outputs().size(); ++i) {
-    auto scalar_type = outputs[i].scalar_type();
-    auto sizes = outputs[i].sizes();
-    auto type =
-        torch::jit::CompleteTensorType::create(scalar_type, at::kCPU, sizes);
-    output_values[i]->setType(type);
-  }
-  return retval;
-}
-
 } // namespace script
 } // namespace jit
 } // namespace torch
